moved computeVersion logic into bump-minor
